+++ lapce-data/src/command.rs
@@ -11,6 +11,7 @@ use lapce_core::{
         EditCommand, FocusCommand, MotionModeCommand, MoveCommand,
         MultiSelectionCommand,
     },
+    movement::LineCol,
     syntax::Syntax,
 };
 use lapce_rpc::{
@@ -39,7 +40,7 @@ use crate::{
         SplitContent,
     },
     document::BufferContent,
-    editor::{EditorLocation, EditorPosition, Line, LineCol},
+    editor::{EditorLocation, EditorPosition, Line},
     images,
     keypress::{KeyMap, KeyPress},
     markdown::Content,
@@ -222,6 +223,10 @@ pub enum LapceWorkbenchCommand {
     #[strum(serialize = "reveal_active_file_in_file_explorer")]
     #[strum(message = "Reveal Active File in File Explorer")]
     RevealActiveFileInFileExplorer,
+
+    #[strum(serialize = "reveal_active_file_in_file_tree")]
+    #[strum(message = "Reveal Active File in File Tree")]
+    RevealActiveFileInFileTree,
 
     #[strum(serialize = "change_color_theme")]
     #[strum(message = "Change Color Theme")]