+++ lapce-proxy/src/lib.rs
@@ -1,6 +1,7 @@
 #![allow(clippy::manual_clamp)]
 
 pub mod buffer;
+pub mod cli;
 pub mod dispatch;
 pub mod plugin;
 pub mod terminal;
@@ -18,8 +19,8 @@ use lapce_rpc::{
 
 pub fn mainloop(
     writer_tx: Sender<RpcMessage<CoreRequest, CoreNotification, ProxyResponse>>,
-    writer_rx: Receiver<RpcMessage<CoreRequest, CoreNotification, ProxyResponse>>,
-    reader_tx: Sender<RpcMessage<ProxyRequest, ProxyNotification, CoreResponse>>,
+    _: Receiver<RpcMessage<CoreRequest, CoreNotification, ProxyResponse>>,
+    _: Sender<RpcMessage<ProxyRequest, ProxyNotification, CoreResponse>>,
     reader_rx: Receiver<RpcMessage<ProxyRequest, ProxyNotification, CoreResponse>>,
 ) {
     let core_rpc = CoreRpcHandler::new();